# docs/接口文档/学习通接口正式交付包/03-联调资产/学习通OpenAPI草案.yaml
openapi: 3.0.3
info:
  title: 学习通 API Draft
  version: 1.1.0
  description: >-
    Draft OpenAPI specification generated from the Study Library requirement,
    interface, DTO, and status-flow documents.
servers:
  - url: /
    description: Unified API root; paths already include full `/api/app/v1` or `/api/admin/v1` prefixes
...
paths:
  /api/app/v1/auth/login:
    post:
      tags: [app-auth]
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppLoginRequestDTO'
      responses:
        '200':
          description: Login response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppLoginResponseEnvelope'
...
  /api/admin/v1/auth/login:
    post:
      tags: [admin-auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginRequestDTO'
      responses:
        '200':
          description: Admin login response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminLoginResponseEnvelope'
...
  /api/admin/v1/home/courses:
    get:
      tags: [admin-config]
      summary: Get home course configs
      parameters:
        - name: sectionType
          in: query
          schema:
            type: string
            enum: [recommend, hot, latest]
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Home course config page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeCourseConfigPageEnvelope'
    post:
      tags: [admin-config]
      summary: Create home course config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HomeCourseConfigSaveRequestDTO'
      responses:
        '200':
          description: Created config id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigIdResponseEnvelope'
  /api/admin/v1/home/courses/{configId}:
    put:
      tags: [admin-config]
      summary: Update home course config
      parameters:
        - name: configId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HomeCourseConfigUpdateRequestDTO'
      responses:
        '200':
          description: Update result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatedAtEnvelope'
    delete:
      tags: [admin-config]
      summary: Delete home course config
      parameters:
        - name: configId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccessEnvelope'
...
  /api/admin/v1/courses/{courseId}/status:
    patch:
      tags: [admin-course]
      summary: Update course publish status
      parameters:
        - $ref: '#/components/parameters/CourseId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [draft, pending, published, offline]
      responses:
        '200':
          description: Status result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseStatusEnvelope'
  /api/admin/v1/course-evaluations:
    get:
      tags: [admin-course]
      summary: Get course evaluation list
      parameters:
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - name: courseId
          in: query
          schema:
            type: integer
            format: int64
        - name: score
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Evaluation page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCourseEvaluationPageEnvelope'
  /api/admin/v1/course-evaluations/{evaluationId}/status:
    patch:
      tags: [admin-course]
      summary: Update course evaluation status
      parameters:
        - name: evaluationId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [visible, hidden, deleted]
      responses:
        '200':
          description: Update result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatedAtEnvelope'
...
  /api/admin/v1/contents:
    get:
      tags: [admin-message]
      summary: Get content list
      parameters:
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - name: contentType
          in: query
          schema:
            type: string
        - name: auditStatus
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Content page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminContentPageEnvelope'
    post:
      tags: [admin-message]
      summary: Create content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminContentSaveRequestDTO'
      responses:
        '200':
          description: Created content id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentIdResponseEnvelope'
  /api/admin/v1/contents/{contentId}:
    get:
      tags: [admin-message]
      summary: Get content detail
      parameters:
        - $ref: '#/components/parameters/ContentId'
      responses:
        '200':
          description: Content detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminContentDetailEnvelope'
    put:
      tags: [admin-message]
      summary: Update content
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminContentSaveRequestDTO'
      responses:
        '200':
          description: Update result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatedAtEnvelope'
    delete:
      tags: [admin-message]
      summary: Delete content
      parameters:
        - $ref: '#/components/parameters/ContentId'
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccessEnvelope'
  /api/admin/v1/contents/{contentId}/audit:
    post:
      tags: [admin-message]
      summary: Audit content
      parameters:
        - $ref: '#/components/parameters/ContentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContentAuditRequestDTO'
      responses:
        '200':
          description: Audit result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContentAuditEnvelope'
...
  /api/admin/v1/messages/{messageId}:
    delete:
      tags: [admin-message]
      summary: Delete message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Delete result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimpleSuccessEnvelope'
...
  /api/admin/v1/message-templates:
    get:
      tags: [admin-config]
      summary: Get message templates
      parameters:
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - name: messageType
          in: query
          schema:
            type: string
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          description: Message template page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageTemplatePageEnvelope'
    post:
      tags: [admin-config]
      summary: Save message template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MessageTemplateSaveRequestDTO'
      responses:
        '200':
          description: Created template id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateIdResponseEnvelope'
...
  /api/admin/v1/service-tickets:
    get:
      tags: [admin-message]
      summary: Get service tickets
      parameters:
        - $ref: '#/components/parameters/PageNum'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Keyword'
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          description: Service ticket page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTicketPageEnvelope'
  /api/admin/v1/service-tickets/{ticketId}:
    get:
      tags: [admin-message]
      summary: Get service ticket detail
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Service ticket detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceTicketDetailEnvelope'
  /api/admin/v1/service-tickets/{ticketId}/reply:
    post:
      tags: [admin-message]
      summary: Reply service ticket
      parameters:
        - name: ticketId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [replyContent]
              properties:
                replyContent:
                  type: string
      responses:
        '200':
          description: Reply result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReplyIdResponseEnvelope'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CourseId:
      name: courseId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    ContentId:
      name: contentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    PageNum:
      name: pageNum
      in: query
      schema:
        type: integer
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        default: 10
    Keyword:
      name: keyword
      in: query
      schema:
        type: string
    Status:
      name: status
      in: query
      schema:
        type: string
  schemas:
    BaseEnvelope:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        requestId:
          type: string
        timestamp:
          type: integer
          format: int64
...
    AppLoginRequestDTO:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          maxLength: 30
        password:
          type: string
          maxLength: 18
        rememberMe:
          type: boolean
...
    AppUserInfo:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        avatar:
          type: string
        memberStatus:
          type: string
          enum: [active, expiring, expired, cancelled]
        roleList:
          type: array
          items:
            type: string
...
    SourceInfo:
      type: object
      properties:
        sourceId:
          type: integer
          format: int64
        sourceName:
          type: string
        sourceLogoUrl:
          type: string
        sourceDesc:
          type: string
...
    AppLoginResponse:
      type: object
      properties:
        token:
          type: string
        refreshToken:
          type: string
        userInfo:
          $ref: '#/components/schemas/AppUserInfo'
...
    AdminLoginRequestDTO:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...
    AdminLoginResponse:
      type: object
      properties:
        token:
          type: string
        adminInfo:
          $ref: '#/components/schemas/AdminInfo'
...
    MessageTemplateSaveRequestDTO:
      type: object
      required: [templateName, messageType, contentTemplate, channelList]
      properties:
        templateName:
          type: string
        messageType:
          type: string
        titleTemplate:
          type: string
        contentTemplate:
          type: string
        channelList:
          type: array
          items:
            type: string
            enum: [site_message, sms, email, push]
...
    AppLoginResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AppLoginResponse'
...
    AdminLoginResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/AdminLoginResponse'
...
    TemplateIdResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                templateId:
                  type: integer
                  format: int64
...
    MessageTemplateItem:
      type: object
      properties:
        templateId:
          type: integer
          format: int64
        templateName:
          type: string
        channelList:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, enabled, disabled]
    MessageTemplatePageEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                pageNum:
                  type: integer
                pageSize:
                  type: integer
                total:
                  type: integer
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/MessageTemplateItem'
...
    SimpleSuccessEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                success:
                  type: boolean
...
    ServiceMessageItem:
      type: object
      properties:
        messageId:
          type: integer
          format: int64
        senderType:
          type: string
        messageText:
          type: string
        sendTime:
          type: string
...
    ServiceTicketItem:
      type: object
      properties:
        ticketId:
          type: integer
          format: int64
        userName:
          type: string
        latestMessage:
          type: string
        status:
          type: string
    ServiceTicketPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                pageNum:
                  type: integer
                pageSize:
                  type: integer
                total:
                  type: integer
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/ServiceTicketItem'
    ReplyIdResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                replyId:
                  type: integer
                  format: int64
                replyTime:
                  type: string
...
    UpdatedAtEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                updatedAt:
                  type: string
...
    HomeCourseConfigSaveRequestDTO:
      type: object
      required: [sectionType, courseId, courseName]
      properties:
        sectionType:
          type: string
          enum: [recommend, hot, latest]
        courseId:
          type: integer
          format: int64
        courseName:
          type: string
        summary:
          type: string
        hotScore:
          type: integer
        sortNo:
          type: integer
    HomeCourseConfigUpdateRequestDTO:
      type: object
      properties:
        summary:
          type: string
        hotScore:
          type: integer
        sortNo:
          type: integer
        displayStatus:
          type: string
          enum: [visible, hidden]
    HomeCourseConfigItem:
      type: object
      properties:
        configId:
          type: integer
          format: int64
        sectionType:
          type: string
        courseName:
          type: string
        hotScore:
          type: integer
    HomeCourseConfigPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                pageNum:
                  type: integer
                pageSize:
                  type: integer
                total:
                  type: integer
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/HomeCourseConfigItem'
    ConfigIdResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                configId:
                  type: integer
                  format: int64
...
    CourseStatusEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                status:
                  type: string
                  enum: [draft, pending, published, offline]
    AdminCourseEvaluationItem:
      type: object
      properties:
        evaluationId:
          type: integer
          format: int64
        courseName:
          type: string
        score:
          type: integer
        status:
          type: string
    AdminCourseEvaluationPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                pageNum:
                  type: integer
                pageSize:
                  type: integer
                total:
                  type: integer
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/AdminCourseEvaluationItem'
...
    AdminContentSaveRequestDTO:
      type: object
      required: [contentType, sourceName, contentText]
      properties:
        contentType:
          type: string
        sourceName:
          type: string
        contentText:
          type: string
        imageList:
          type: array
          items:
            type: string
    AdminContentItem:
      type: object
      properties:
        contentId:
          type: integer
          format: int64
        contentType:
          type: string
        authorName:
          type: string
        auditStatus:
          type: string
    AdminContentPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                pageNum:
                  type: integer
                pageSize:
                  type: integer
                total:
                  type: integer
                list:
                  type: array
                  items:
                    $ref: '#/components/schemas/AdminContentItem'
    ContentIdResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                contentId:
                  type: integer
                  format: int64
...
    ServiceTicketDetailEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                ticketInfo:
                  $ref: '#/components/schemas/ServiceTicketItem'
                messageList:
                  type: array
                  items:
                    $ref: '#/components/schemas/ServiceMessageItem'
    AdminContentDetailEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                contentId:
                  type: integer
                  format: int64
                sourceInfo:
                  $ref: '#/components/schemas/SourceInfo'
                contentText:
                  type: string
                auditStatus:
                  type: string
    ContentAuditRequestDTO:
      type: object
      required: [auditStatus]
      properties:
        auditStatus:
          type: string
          enum: [pending, approved, rejected]
        auditRemark:
          type: string
    ContentAuditEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data:
              type: object
              properties:
                auditStatus:
                  type: string
                  enum: [pending, approved, rejected]
...